restructure mysql tests
